requirements.txt
streamlit
numpy
opencv-python
Pillow
pytesseract
scikit-image